Calendar controls working with mock data.
