+++ app/javascript/react/types/fixedStream.ts
@@ -36,18 +36,6 @@ interface FixedStreamShortInfo extends FixedStreamStationInfo {
   lastMeasurementDateLabel: string | undefined;
 }
 
-interface CalendarCellData {
-  date: string;
-  dayNumber: string;
-  value: number | null;
-}
-
-interface CalendarMonthlyData {
-  monthName: string;
-  dayNamesHeader: string[];
-  weeks: CalendarCellData[][];
-}
-
 export type {
   StreamUpdate,
   DataSource,
@@ -55,7 +43,5 @@ export type {
   StreamDailyAverage,
   FixedStream,
   FixedStreamShortInfo,
-  CalendarCellData,
-  CalendarMonthlyData,
-  Measurement
+  Measurement,
 };
